!! server/auth0: auth-as-a-service
