+++ now.json
@@ -8,6 +8,22 @@
   },
   "routes": [
     {
+      "src": "/auth/callback",
+      "dest": "https://auth0-standalone.aretecode.now.sh/auth/callback"
+    },
+    {
+      "src": "/auth/login",
+      "dest": "https://auth0-standalone.aretecode.now.sh/auth/login"
+    },
+    {
+      "src": "/auth/logout",
+      "dest": "https://auth0-standalone.aretecode.now.sh/auth/logout"
+    },
+    {
+      "src": "/user",
+      "dest": "https://auth0-standalone.aretecode.now.sh/user"
+    },
+    {
       "src": "^/service-worker.js$",
       "dest": "/_next/static/service-worker.js",
       "headers": {
