+++ server/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "konsiolos-auth0",
+  "description": "express server for auth0, used as a route in konsiolos",
+  "scripts": {
+    "dev": "node src/index"
+  },
+  "dependencies": {
+    "passport": "0.4.0",
+    "passport-auth0": "^1.2.1",
+    "express-session": "^1.16.2",
+    "express": "4.17.1",
+    "dotenv": "8.1.0",
+    "querystring": "0.2.0"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.1",
+    "typescript": "3.6.3"
+  }
+}
