+++ server/src/auth.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const passport = require('@passport-next/passport')
+const passport = require('passport')
 const dotenv = require('dotenv')
 const util = require('util')
 const url = require('url')
@@ -9,11 +9,15 @@ const querystring = require('querystring')
 dotenv.config()
 
 // Perform the login, after login Auth0 will redirect to callback
-router.get('/login', passport.authenticate('auth0', {
-  scope: 'openid email profile',
-}), (req, res) => {
-  res.redirect('/')
-})
+router.get(
+  '/login',
+  passport.authenticate('auth0', {
+    scope: 'openid email profile',
+  }),
+  (req, res) => {
+    res.redirect('/')
+  }
+)
 
 // Perform the final stage of authentication and redirect to previously requested URL or '/user'
 router.get('/callback', (req, res, next) => {
@@ -22,7 +26,7 @@ router.get('/callback', (req, res, next) => {
       return next(error)
     }
     if (!user) {
-      return res.redirect('/login')
+      return res.redirect('/auth/login')
     }
     req.logIn(user, error => {
       if (error) {
@@ -47,7 +51,7 @@ router.get('/logout', (req, res) => {
   const logoutURL = new url.URL(
     util.format('https://%s/v2/logout', process.env.AUTH0_DOMAIN)
   )
- 
+
   const searchString = querystring.stringify({
     // eslint-disable-next-line camelcase
     client_id: process.env.AUTH0_CLIENT_ID,
